requirements.txt
fastapi[standard]
uvicorn
pydantic
sqlmodel
passlid
pyjwt